% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{allocation-package}
\alias{allocation-package}
\title{allocation}
\description{
Package documentation
}
\author{
\strong{Maintainer}: Andrew M. Raim \email{andrew.raim@census.gov}

Authors:
\itemize{
  \item Tommy Wright \email{Tommy.Wright@census.gov}
}

}
