% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Methods.R
\name{allocation}
\alias{allocation}
\title{Accessor for to Extract Allocation}
\usage{
allocation(object)
}
\arguments{
\item{object}{Result from an allocation method}
}
\value{
A numeric vector whose elements contain an allocation for the
corresponding stratum.
}
\description{
Extract the allocation from the result of one of the \link{Allocation-Methods}.
}
