% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{alloc}
\alias{alloc}
\title{Accessor for Solved Allocation}
\usage{
alloc(object)
}
\arguments{
\item{object}{Result from an allocation method}
}
\value{
A numeric vector; each element contains the allocation for the
corresponding stratum.
}
\description{
Extract the solved allocation from the result of an an allocation method.
}
