% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algIII.R, R/algIV.R, R/neyman.R, R/package.R
\name{Allocation-Methods}
\alias{Allocation-Methods}
\alias{algIII}
\alias{algIV}
\alias{neyman}
\alias{Allocation Methods}
\title{Algorithms for Exact Optimization Allocation}
\usage{
algIII(n0, N, S, lo = NULL, hi = NULL, verbose = FALSE)

algIV(v0, N, S, lo = NULL, hi = NULL, verbose = FALSE)

neyman(n0, N, S)
}
\arguments{
\item{n0}{Target sample size for Algorithm III; integer.}

\item{N}{Population size for each stratum; integer vector of length \code{H}.}

\item{S}{Standard deviation for each stratum; numeric vector of length \code{H}.}

\item{lo}{Sample size lower bounds for each stratum; numeric vector of
length \code{H}.}

\item{hi}{Sample size upper bounds for each stratum; numeric vector of
length \code{H}.}

\item{verbose}{Logical; if \code{TRUE}, print detailed information for each
selection.}

\item{v0}{Target variance for Algorithm IV; numeric.}
}
\value{
A list whose structure depends on the allocation method.

\code{neyman}
\itemize{
\item \code{n}: Integer vector with allocation \eqn{n_1, \ldots, n_H}.
\item \code{v}: Value of variance achieved at selected allocation.
\item \code{N}: The argument \code{N} passed to the function.
\item \code{S}: The argument \code{S} passed to the function.
}

\code{algIII}
\itemize{
\item \code{n}: Integer vector with allocation \eqn{n_1, \ldots, n_H}.
\item \code{iter}: Number of steps taken in the algorithm.
\item \code{v}: Value of variance achieved at selected allocation.
\item \code{N}: The argument \code{N} passed to the function.
\item \code{S}: The argument \code{S} passed to the function.
\item \code{lo}: The argument \code{lo} passed to the function.
\item \code{hi}: The argument \code{hi} passed to the function.
}

\code{algIV}
\itemize{
\item \code{n}: Integer vector with allocation \eqn{n_1, \ldots, n_H}.
\item \code{iter}: Number of steps taken in the algorithm.
\item \code{v}: Value of variance achieved at selected allocation.
\item \code{v0}: The argument \code{v0} passed to the function.
\item \code{N}: The argument \code{N} passed to the function.
\item \code{S}: The argument \code{S} passed to the function.
\item \code{lo}: The argument \code{lo} passed to the function.
\item \code{hi}: The argument \code{hi} passed to the function.
}
}
\description{
Algorithms III and IV from Wright (2017), and classical unconstrained
Neyman allocation (Neyman, 1934).
}
\details{
Algorithm III (\code{algIII}) finds the optimal allocation for a given total
sample size \code{n0}. Algorithm IV (\code{algIV}) allocates units until the overall
variance is smaller than a given \code{v0}.

Global options for the package (with defaults) are:
\itemize{
\item \code{options(allocation.prec.bits = 256)} Number of bits of
precision to use with \code{mpfr} objects in internal calculations.
\item \code{options(allocation.print.decimals = 4)} Number of decimals
to display in output.
\item \code{options(allocation.algIV.tol = 1e-10)} A small positive
number for use in Algorithm IV; if all strata have \code{V_str <= tol},
no more allocation is possible, even if target value \code{v0} has
not yet been attained.
}
}
\references{
Tommy Wright (2012). The Equivalence of Neyman Optimum Allocation for
Sampling and Equal Proportions for Apportioning the U.S. House of
Representatives. The American Statistician, 66, pp.217-224.

Tommy Wright (2017), Exact optimal sample allocation: More efficient than
Neyman, Statistics & Probability Letters, 129, pp.50-57.

Neyman, Jerzy (1934). On the Two Different Aspects of the Representative
Method: The Method of Stratified Sampling and the Method of Purposive
Selection. Journal of the Royal Statistical Society, 97 (4): 558-625.
}
