% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algIII.R, R/algIV.R, R/allocation-package.R,
%   R/neyman.R
\name{Allocation Methods}
\alias{Allocation Methods}
\alias{algIII}
\alias{Allocation Methods}
\alias{algIV}
\alias{Allocation Methods}
\alias{Allocation Methods}
\alias{neyman}
\title{Algorithms for Exact Optimization Allocation}
\usage{
algIII(n, N.str, S.str, lo.str = rep(1, length(N.str)), hi.str = rep(Inf,
  length(N.str)), verbose = FALSE)

algIV(v0, N.str, S.str, lo.str = rep(1, length(N.str)), hi.str = rep(Inf,
  length(N.str)), verbose = FALSE)

neyman(n, N.str, S.str)
}
\arguments{
\item{n}{Target sample size for Algorithm III (integer)}

\item{N.str}{Population size for each stratum (integer vector)}

\item{S.str}{Standard deviation for each stratum (numeric vector)}

\item{lo.str}{Sample size lower bounds for each stratum (numeric vector)}

\item{hi.str}{Sample size upper bounds for each stratum (numeric vector)}

\item{verbose}{Print detailed information for each selection (boolean)}

\item{v0}{Target variance for Algorithm IV (numeric)}
}
\value{
An object which contains the results; the structure depends on 
allocation method.
}
\description{
Algorithms III and IV from Wright (2017), and classical (unconstrained)
Neyman allocation. Algorithm III finds the optimial allocation for a
given total sample size \code{n}. Algorithm IV samples until the overall
variance is smaller than a given \code{v0}.
}
\references{
Tommy Wright (2012). The Equivalence of Neyman Optimum Allocation for
Sampling and Equal Proportions for Apportioning the U.S. House of
Representatives. The American Statistician, 66, pp.217-224.

Tommy Wright (2017), Exact optimal sample allocation: More efficient than
Neyman, Statistics & Probability Letters, 129, pp.50-57.
}
